implement and test the read pair build

+++ SR_Common/SR_BamMemPool.h
@@ -63,7 +63,7 @@ struct SR_BamBuff
 
 struct SR_BamMemPool
 {
-    unsigned short numBuffs;
+    unsigned int numBuffs;
 
     unsigned int buffCapacity;
